refactor: organize icon stories into individual files per component
Improve story organization by separating each icon into its own story file in src/stories/icons/:
- Create individual .stories.tsx files for Check, Error, Warning, Plus, HamburgerMenu, X, and all Chevrons
- Extract shared IconVariations component to src/stories/icons/IconVariations.tsx for reuse
- Update Icons.stories.tsx to retain only overview/showcase stories (AllDefault, AllBordered, AllInverted, AllBorderedInverted, Sizes, SpinnerAnimation, ChevronRotation, ChevronAdditiveRotation)
- Name all icon story files with .stories extension so Storybook's file discovery picks them up

Result: Icons now appear as a group in Storybook sidebar with individual stories for each icon, making them easier to find and maintain.
