+++ src/stories/Icons.stories.tsx
@@ -120,61 +120,6 @@ export const Sizes: StoryObj = {
   ),
 };
 
-export const SpinnerAnimation: StoryObj = {
-  render: () => (
-    <div style={{ display: "flex", gap: "3rem" }}>
-      <div style={{ textAlign: "center" }}>
-        <Icons.Spinner />
-        <p style={{ marginTop: "1rem" }}>Default</p>
-      </div>
-      <div style={{ textAlign: "center" }}>
-        <Icons.Spinner bordered />
-        <p style={{ marginTop: "1rem" }}>Bordered</p>
-      </div>
-      <div
-        style={{
-          textAlign: "center",
-          padding: "1rem",
-          backgroundColor: "#f0f0f0",
-        }}
-      >
-        <Icons.Spinner inverted />
-        <p style={{ marginTop: "1rem" }}>Inverted</p>
-      </div>
-    </div>
-  ),
-};
-
-export const ChevronRotation: StoryObj = {
-  render: () => (
-    <div
-      style={{
-        display: "flex",
-        gap: "2rem",
-        alignItems: "center",
-        justifyContent: "center",
-      }}
-    >
-      <div style={{ textAlign: "center" }}>
-        <Icons.ChevronUp />
-        <p>ChevronUp</p>
-      </div>
-      <div style={{ textAlign: "center" }}>
-        <Icons.ChevronRight />
-        <p>ChevronRight</p>
-      </div>
-      <div style={{ textAlign: "center" }}>
-        <Icons.ChevronDown />
-        <p>ChevronDown</p>
-      </div>
-      <div style={{ textAlign: "center" }}>
-        <Icons.ChevronLeft />
-        <p>ChevronLeft</p>
-      </div>
-    </div>
-  ),
-};
-
 export const ChevronAdditiveRotation: StoryObj = {
   render: () => (
     <div
@@ -199,83 +144,4 @@ export const ChevronAdditiveRotation: StoryObj = {
       </div>
     </div>
   ),
-};
-
-function IconVariations({
-  Component,
-}: {
-  Component: React.ComponentType<any>;
-}) {
-  return (
-    <div style={{ display: "flex", gap: "2rem" }}>
-      <div style={{ textAlign: "center" }}>
-        <Component />
-        <p>Default</p>
-      </div>
-      <div style={{ textAlign: "center" }}>
-        <Component bordered />
-        <p>Bordered</p>
-      </div>
-      <div
-        style={{
-          textAlign: "center",
-          padding: "1rem",
-          backgroundColor: "#f0f0f0",
-        }}
-      >
-        <Component inverted />
-        <p>Inverted</p>
-      </div>
-      <div
-        style={{
-          textAlign: "center",
-          padding: "1rem",
-          backgroundColor: "#f0f0f0",
-        }}
-      >
-        <Component bordered inverted />
-        <p>Bordered + Inverted</p>
-      </div>
-    </div>
-  );
-}
-
-export const Check: StoryObj = {
-  render: () => <IconVariations Component={Icons.Check} />,
-};
-
-export const Error: StoryObj = {
-  render: () => <IconVariations Component={Icons.Error} />,
-};
-
-export const Warning: StoryObj = {
-  render: () => <IconVariations Component={Icons.Warning} />,
-};
-
-export const Plus: StoryObj = {
-  render: () => <IconVariations Component={Icons.Plus} />,
-};
-
-export const HamburgerMenu: StoryObj = {
-  render: () => <IconVariations Component={Icons.HamburgerMenu} />,
-};
-
-export const X: StoryObj = {
-  render: () => <IconVariations Component={Icons.X} />,
-};
-
-export const ChevronDown: StoryObj = {
-  render: () => <IconVariations Component={Icons.ChevronDown} />,
-};
-
-export const ChevronUp: StoryObj = {
-  render: () => <IconVariations Component={Icons.ChevronUp} />,
-};
-
-export const ChevronLeft: StoryObj = {
-  render: () => <IconVariations Component={Icons.ChevronLeft} />,
-};
-
-export const ChevronRight: StoryObj = {
-  render: () => <IconVariations Component={Icons.ChevronRight} />,
 };
